dom/tests/mochitest/pointerlock/file_loseFocusWindow.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_exitMouseLockOnLoseFocus.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <script type="application/javascript" src="pointerlock_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>
  <div id="div"></div>
  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * If element has the pointer locked and window loses focus
       * pointer should be unlocked
       */

      SimpleTest.waitForExplicitFinish();

      var div = document.getElementById("div")
        , newWindow = null
        , looseFocusUnlock = false
        , pointerLockLost = false;

      function runTests () {
        ok(looseFocusUnlock, "Pointer should be unlocked if window " +
            "loses focus");
      }

      document.addEventListener("mozpointerlockchange", function (e) {
        if (document.mozPointerLockElement === div) {
          divPointerLocked = true;
          newWindow = window.open(); // Open new window to blur current one
        }
        else {
          looseFocusUnlock = true;
          newWindow && newWindow.close();
          document.mozCancelFullScreen();
        }
      }, false);

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreenElement === div) {
          div.mozRequestPointerLock();
        }
        else {
          runTests();
          SimpleTest.finish();
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          div.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
